pyproject.toml
[project]
name = "amantra-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi>=0.135.1",
    "orjson>=3.11.7",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.1",
]

[dependency-groups]
dev = [
    "uvicorn>=0.42.0",
]
